pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "align-mamba"
version = "0.1.0"
description = "State Capacity Limits in Selective SSMs"
requires-python = ">=3.10"

dependencies = [
    "torch>=2.3.0",
    "numpy<2",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "tokenizers>=0.19",
    "datasets>=2.19",
    "mamba-ssm>=2.0",
    "causal-conv1d>=1.2",
    "flash-attn>=2.5",
]

[project.scripts]
align-train = "align_mamba.scripts.train:main"
align-eval = "align_mamba.scripts.evaluate:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["align_mamba*"]
